oauth and jwt

+++ app/routes/events.py
@@ -1,12 +1,14 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from beanie import PydanticObjectId
-from fastapi import APIRouter, HTTPException, status
-from ..database.connection import Database
-
-from ..models.events import Event, EventUpdate
 from typing import List
+
+from beanie import PydanticObjectId
+from fastapi import APIRouter, Depends, HTTPException, status
+
+from ..auth.authenticate import authenticate  # event 라우트에 의존성 주입
+from ..database.connection import Database
+from ..models.events import Event, EventUpdate
 
 event_router = APIRouter(tags=["Event"])
 event_database = Database(Event)
@@ -29,14 +31,22 @@ async def get_event(id: PydanticObjectId) -> Event:
 
 # create event
 @event_router.post("/new")
-async def create_event(body: Event) -> dict:
+async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
+    body.creator = user
     await event_database.save(body)
     return {"message": "Event created successfully"}
 
 
 # update event
 @event_router.put("/{id}")
-async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
+async def update_event(
+    id: PydanticObjectId, body: EventUpdate,
+    user: str = Depends(authenticate)) -> Event:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
+                            detail="Operation not permitted")
+
     event = await event_database.update(id, body)
     if not event:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
@@ -46,7 +56,13 @@ async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
 
 # delete event router
 @event_router.delete("/{id}")
-async def delete_event(id: PydanticObjectId) -> dict:
+async def delete_event(
+    id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
+                            detail="Operation not permitted")
+
     event = await event_database.delete(id)
     if not event:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
@@ -54,8 +70,8 @@ async def delete_event(id: PydanticObjectId) -> dict:
     return {"message": "Event deleted successfully"}
 
 
-# delete all events
-@event_router.delete("/")
-async def delete_all_events() -> dict:
-    await event_database.delete_all()
-    return {"message": "All events deleted"}
+##  delete all events
+#  @event_router.delete("/")
+#  async def delete_all_events(user: str = Depends(authenticate)) -> dict:
+#  await event_database.delete_all()
+#  return {"message": "All events deleted"}
